cron/requirements.txt
apscheduler==3.10.4
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
plaid-python==39.0.0
anthropic==0.34.2
numpy==2.2.4
scikit-learn==1.6.1
joblib==1.4.2
